Split off compose
