+++ docs/viewer/test.ts
@@ -1,5 +1,5 @@
 import { Ifc5FileJson } from "../../schema/out/@typespec/json-schema/ts/ifc5file";
-import { compose } from "./render";
+import { compose } from "./compose";
 let fs = require("fs");
 
 let helloWallFileName = "../../Hello Wall/hello-wall.ifcx";
@@ -8,3 +8,5 @@ let helloWallJSON = JSON.parse(fs.readFileSync(helloWallFileName).toString());
 console.log(helloWallJSON);
 
 let composed = compose([helloWallJSON] as Ifc5FileJson[]);
+
+console.log(composed);
